Fix DataHubTest,springify UpgradeProjectTest,remove HubProjectImplTest
