+++ marklogic-data-hub/src/test/java/com/marklogic/hub/impl/UpgradeProjectTest.java
@@ -1,10 +1,20 @@
 package com.marklogic.hub.impl;
 
+import com.marklogic.hub.ApplicationConfig;
 import com.marklogic.hub.HubConfig;
+import com.marklogic.hub.HubTestBase;
 import org.apache.commons.io.FileUtils;
+import org.junit.jupiter.api.AfterEach;
+import org.junit.jupiter.api.Assertions;
+import org.junit.jupiter.api.Assumptions;
 import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.extension.ExtendWith;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.test.context.ContextConfiguration;
+import org.springframework.test.context.junit.jupiter.SpringExtension;
 
 import java.io.File;
+import java.io.IOException;
 import java.nio.file.Paths;
 import java.util.HashMap;
 import java.util.Map;
@@ -16,7 +26,23 @@ import static org.junit.jupiter.api.Assertions.assertEquals;
  * src/test/resources/upgrade-projects into the build directory (a non-version-controlled area) where it
  * can then be upgraded and verified.
  */
-public class UpgradeProjectTest {
+@ExtendWith(SpringExtension.class)
+@ContextConfiguration(classes = ApplicationConfig.class)
+public class UpgradeProjectTest extends HubTestBase {
+    @Autowired
+    private FlowManagerImpl flowManager;
+
+    @Autowired
+    private HubProjectImpl hubProject;
+
+    @Autowired
+    private Versions versions;
+
+    @AfterEach
+    public void cleanup() {
+        deleteProjectDir();
+    }
+
 
     @Test
     public void upgrade43xToCurrentVersion() throws Exception {
@@ -24,8 +50,6 @@ public class UpgradeProjectTest {
         final File projectDir = Paths.get(projectPath).toFile();
         FileUtils.deleteDirectory(projectDir);
         FileUtils.copyDirectory(Paths.get("src/test/resources/upgrade-projects/dhf43x").toFile(), projectDir);
-
-        HubProjectImpl hubProject = new HubProjectImpl();
         hubProject.createProject(projectPath);
         hubProject.init(createMap());
         hubProject.upgradeProject();
@@ -44,5 +68,18 @@ public class UpgradeProjectTest {
 
     private void verifyDirContents(File dir, int expectedCount) {
         assertEquals(expectedCount, dir.listFiles().length);
+    }
+
+    @Test
+    public void testUpgradeTo510MappingStep() throws IOException {
+        createProjectDir();
+        adminHubConfig.createProject(PROJECT_PATH);
+        Assumptions.assumingThat(versions.isVersionCompatibleWithES(), () -> {
+            FileUtils.copyFileToDirectory(getResourceFile("mapping-test/flows/CustomerXML.flow.json"), adminHubConfig.getFlowsDir().toFile());
+            FileUtils.copyDirectory(getResourceFile("flow-runner-test/flows"), adminHubConfig.getFlowsDir().toFile());
+            project.upgradeFlows();
+            Assertions.assertEquals("entity-services-mapping", flowManager.getFlow("testFlow").getStep("6").getStepDefinitionName());
+            Assertions.assertEquals("entity-services-mapping", flowManager.getFlow("CustomerXML").getStep("2").getStepDefinitionName());
+        });
     }
 }
